requirements.txt
requests==2.21.0
python-dotenv==0.10.2
python-telegram-bot==11.1.0
vk-api==11.3.0
